feat: refactoring the admin panel, with new paths for better management
